extracted imageloader into its own file
